+++ x/incentives/keeper/gauge_test.go
@@ -8,22 +8,60 @@ import (
 	lockuptypes "github.com/osmosis-labs/osmosis/x/lockup/types"
 )
 
+// TestDistribute tests that when the distribute command is executed on
+// a provided gauge,
+func (suite *KeeperTestSuite) TestDistribute() {
+	twoLockupUser := userLocks{
+		lockDurations: []time.Duration{time.Second, 2 * time.Second},
+		lockAmounts:   []sdk.Coins{defaultLPTokens, defaultLPTokens},
+	}
+	defaultGauge := perpGaugeDesc{
+		lockDenom:    defaultLPDenom,
+		lockDuration: defaultLockDuration,
+		rewardAmount: sdk.Coins{sdk.NewInt64Coin(defaultRewardDenom, 3000)},
+	}
+	oneKRewardCoins := sdk.Coins{sdk.NewInt64Coin(defaultRewardDenom, 1000)}
+	twoKRewardCoins := oneKRewardCoins.Add(oneKRewardCoins...)
+	tests := []struct {
+		users           []userLocks
+		gauges          []perpGaugeDesc
+		expectedRewards []sdk.Coins
+	}{
+		{
+			users:           []userLocks{oneLockupUser, twoLockupUser},
+			gauges:          []perpGaugeDesc{defaultGauge},
+			expectedRewards: []sdk.Coins{oneKRewardCoins, twoKRewardCoins},
+		},
+	}
+	for _, tc := range tests {
+		suite.SetupTest()
+		gauges := suite.SetupGauges(tc.gauges)
+		addrs := suite.SetupUserLocks(tc.users)
+		for _, g := range gauges {
+			suite.app.IncentivesKeeper.Distribute(suite.ctx, g)
+		}
+		// Check expected rewards
+		for i, addr := range addrs {
+			bal := suite.app.BankKeeper.GetAllBalances(suite.ctx, addr)
+			suite.Require().Equal(bal.String(), tc.expectedRewards[i].String())
+		}
+	}
+}
+
 func (suite *KeeperTestSuite) TestInvalidDurationGaugeCreationValidation() {
 	suite.SetupTest()
 
-	addr := sdk.AccAddress([]byte("addr1---------------"))
-	coins := sdk.Coins{sdk.NewInt64Coin("stake", 10)}
-	suite.app.BankKeeper.SetBalances(suite.ctx, addr, coins)
+	addrs := suite.SetupManyLocks(1, defaultLiquidTokens, defaultLPTokens, defaultLockDuration)
 	distrTo := lockuptypes.QueryCondition{
 		LockQueryType: lockuptypes.ByDuration,
-		Denom:         "lptoken",
-		Duration:      time.Second / 2, // 0.5 second
+		Denom:         defaultLPDenom,
+		Duration:      defaultLockDuration / 2, // 0.5 second, invalid duration
 	}
-	_, err := suite.app.IncentivesKeeper.CreateGauge(suite.ctx, false, addr, coins, distrTo, time.Time{}, 1)
+	_, err := suite.app.IncentivesKeeper.CreateGauge(suite.ctx, false, addrs[0], defaultLiquidTokens, distrTo, time.Time{}, 1)
 	suite.Require().Error(err)
 
-	distrTo.Duration = time.Second
-	_, err = suite.app.IncentivesKeeper.CreateGauge(suite.ctx, false, addr, coins, distrTo, time.Time{}, 1)
+	distrTo.Duration = defaultLockDuration
+	_, err = suite.app.IncentivesKeeper.CreateGauge(suite.ctx, false, addrs[0], defaultLiquidTokens, distrTo, time.Time{}, 1)
 	suite.Require().NoError(err)
 }
 
@@ -110,8 +148,10 @@ func (suite *KeeperTestSuite) TestNonPerpetualGaugeOperations() {
 	gauges := suite.app.IncentivesKeeper.GetNotFinishedGauges(suite.ctx)
 	suite.Require().Len(gauges, 0)
 
-	// setup lock and gauge
-	lockOwner, gaugeID, coins, startTime := suite.SetupLockAndGauge(false)
+	lockOwners := suite.SetupManyLocks(5, defaultLiquidTokens, defaultLPTokens, time.Second)
+	gaugeID, _, coins, startTime := suite.SetupNewGauge(false, sdk.Coins{sdk.NewInt64Coin("stake", 10)})
+	// evenly distributed per lock
+	expectedCoinsPerLock := sdk.Coins{sdk.NewInt64Coin("stake", 2)}
 
 	// check gauges
 	gauges = suite.app.IncentivesKeeper.GetNotFinishedGauges(suite.ctx)
@@ -124,8 +164,8 @@ func (suite *KeeperTestSuite) TestNonPerpetualGaugeOperations() {
 	suite.Require().Equal(gauges[0].StartTime.Unix(), startTime.Unix())
 
 	// check rewards estimation
-	rewardsEst := suite.app.IncentivesKeeper.GetRewardsEst(suite.ctx, lockOwner, []lockuptypes.PeriodLock{}, 100)
-	suite.Require().Equal(coins.String(), rewardsEst.String())
+	rewardsEst := suite.app.IncentivesKeeper.GetRewardsEst(suite.ctx, lockOwners[0], []lockuptypes.PeriodLock{}, 100)
+	suite.Require().Equal(expectedCoinsPerLock.String(), rewardsEst.String())
 
 	// add to gauge
 	addCoins := sdk.Coins{sdk.NewInt64Coin("stake", 200)}
@@ -196,8 +236,7 @@ func (suite *KeeperTestSuite) TestNonPerpetualGaugeOperations() {
 	_, err = suite.app.IncentivesKeeper.GetGaugeByID(suite.ctx, gaugeID+1000)
 	suite.Require().Error(err)
 
-	// check rewards estimation
-	rewardsEst = suite.app.IncentivesKeeper.GetRewardsEst(suite.ctx, lockOwner, []lockuptypes.PeriodLock{}, 100)
+	rewardsEst = suite.app.IncentivesKeeper.GetRewardsEst(suite.ctx, lockOwners[0], []lockuptypes.PeriodLock{}, 100)
 	suite.Require().Equal(sdk.Coins{}, rewardsEst)
 }
 
@@ -399,86 +438,59 @@ func (suite *KeeperTestSuite) TestNoLockNonPerpetualGaugeDistribution() {
 	suite.Require().Equal(gauges[0].String(), expectedGauge.String())
 }
 
-func (suite *KeeperTestSuite) TestNonPerpetualActiveGaugesByDenom() {
-	// test for module get gauges
-	suite.SetupTest()
+func (suite *KeeperTestSuite) TestGaugesByDenom() {
+	// TODO: This is not a good test. We should refactor it to be table driven,
+	// specifying a list of gauges to define, and then the expected result of gauges by denom
+	testGaugeByDenom := func(isPerpetual bool) {
+		// test for module get gauges
+		suite.SetupTest()
 
-	// initial module gauges check
-	gaugeIds := suite.app.IncentivesKeeper.GetAllGaugeIDsByDenom(suite.ctx, "lptoken")
-	suite.Require().Len(gaugeIds, 0)
+		// initial module gauges check
+		gaugeIds := suite.app.IncentivesKeeper.GetAllGaugeIDsByDenom(suite.ctx, "lptoken")
+		suite.Require().Len(gaugeIds, 0)
 
-	// setup lock and gauge
-	_, gaugeID, _, startTime := suite.SetupLockAndGauge(false)
+		// setup lock and gauge
+		_, gaugeID, _, startTime := suite.SetupLockAndGauge(isPerpetual)
 
-	// check gauges
-	gaugeIds = suite.app.IncentivesKeeper.GetAllGaugeIDsByDenom(suite.ctx, "lptoken")
-	suite.Require().Len(gaugeIds, 1)
-	suite.Require().Equal(gaugeIds[0], gaugeID)
+		// check gauges
+		gaugeIds = suite.app.IncentivesKeeper.GetAllGaugeIDsByDenom(suite.ctx, "lptoken")
+		suite.Require().Len(gaugeIds, 1, "perpetual %b", isPerpetual)
+		suite.Require().Equal(gaugeIds[0], gaugeID)
 
-	// start distribution
-	suite.ctx = suite.ctx.WithBlockTime(startTime)
-	gauge, err := suite.app.IncentivesKeeper.GetGaugeByID(suite.ctx, gaugeID)
-	suite.Require().NoError(err)
-	err = suite.app.IncentivesKeeper.BeginDistribution(suite.ctx, *gauge)
-	suite.Require().NoError(err)
+		// start distribution
+		suite.ctx = suite.ctx.WithBlockTime(startTime)
+		gauge, err := suite.app.IncentivesKeeper.GetGaugeByID(suite.ctx, gaugeID)
+		suite.Require().NoError(err)
+		err = suite.app.IncentivesKeeper.BeginDistribution(suite.ctx, *gauge)
+		suite.Require().NoError(err)
 
-	// check gauge ids by denom
-	gaugeIds = suite.app.IncentivesKeeper.GetAllGaugeIDsByDenom(suite.ctx, "lptoken")
-	suite.Require().Len(gaugeIds, 1)
+		// check gauge ids by denom
+		gaugeIds = suite.app.IncentivesKeeper.GetAllGaugeIDsByDenom(suite.ctx, "lptoken")
+		suite.Require().Len(gaugeIds, 1)
 
-	// distribute coins to stakers
-	_, err = suite.app.IncentivesKeeper.Distribute(suite.ctx, *gauge)
-	suite.Require().NoError(err)
+		// check gauge ids by other denom
+		gaugeIds = suite.app.IncentivesKeeper.GetAllGaugeIDsByDenom(suite.ctx, "lpt")
+		suite.Require().Len(gaugeIds, 0)
 
-	// finish distribution
-	err = suite.app.IncentivesKeeper.FinishDistribution(suite.ctx, *gauge)
-	suite.Require().NoError(err)
+		// distribute coins to stakers
+		_, err = suite.app.IncentivesKeeper.Distribute(suite.ctx, *gauge)
+		suite.Require().NoError(err)
 
-	// check gauge ids by denom
-	gaugeIds = suite.app.IncentivesKeeper.GetAllGaugeIDsByDenom(suite.ctx, "lptoken")
-	suite.Require().Len(gaugeIds, 0)
-}
+		// finish distribution for non perpetual gauge
+		if !gauge.IsPerpetual {
+			err = suite.app.IncentivesKeeper.FinishDistribution(suite.ctx, *gauge)
+			suite.Require().NoError(err)
+		}
 
-func (suite *KeeperTestSuite) TestPerpetualActiveGaugesByDenom() {
-	// test for module get gauges
-	suite.SetupTest()
+		expectedNumGauges := 1
+		if !isPerpetual {
+			expectedNumGauges = 0
+		}
+		// check gauge ids by denom
+		gaugeIds = suite.app.IncentivesKeeper.GetAllGaugeIDsByDenom(suite.ctx, "lptoken")
+		suite.Require().Len(gaugeIds, expectedNumGauges)
+	}
 
-	// initial module gauges check
-	gaugeIds := suite.app.IncentivesKeeper.GetAllGaugeIDsByDenom(suite.ctx, "lptoken")
-	suite.Require().Len(gaugeIds, 0)
-
-	// setup lock and gauge
-	_, gaugeID, _, startTime := suite.SetupLockAndGauge(true)
-
-	// check gauges
-	gaugeIds = suite.app.IncentivesKeeper.GetAllGaugeIDsByDenom(suite.ctx, "lptoken")
-	suite.Require().Len(gaugeIds, 1)
-	suite.Require().Equal(gaugeIds[0], gaugeID)
-
-	// start distribution
-	suite.ctx = suite.ctx.WithBlockTime(startTime)
-	gauge, err := suite.app.IncentivesKeeper.GetGaugeByID(suite.ctx, gaugeID)
-	suite.Require().NoError(err)
-	err = suite.app.IncentivesKeeper.BeginDistribution(suite.ctx, *gauge)
-	suite.Require().NoError(err)
-
-	// check gauge ids by denom
-	gaugeIds = suite.app.IncentivesKeeper.GetAllGaugeIDsByDenom(suite.ctx, "lptoken")
-	suite.Require().Len(gaugeIds, 1)
-
-	// check gauge ids by other denom
-	gaugeIds = suite.app.IncentivesKeeper.GetAllGaugeIDsByDenom(suite.ctx, "lpt")
-	suite.Require().Len(gaugeIds, 0)
-
-	// check gauge ids by other denom
-	gaugeIds = suite.app.IncentivesKeeper.GetAllGaugeIDsByDenom(suite.ctx, "token")
-	suite.Require().Len(gaugeIds, 0)
-
-	// distribute coins to stakers
-	_, err = suite.app.IncentivesKeeper.Distribute(suite.ctx, *gauge)
-	suite.Require().NoError(err)
-
-	// check gauge ids by denom
-	gaugeIds = suite.app.IncentivesKeeper.GetAllGaugeIDsByDenom(suite.ctx, "lptoken")
-	suite.Require().Len(gaugeIds, 1)
+	testGaugeByDenom(true)
+	testGaugeByDenom(false)
 }